+++ docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-07-08
+
+- Added migration logic to handle TLS private key and CSR migration.
+
 ## 2026-07-01
 
 - Added a `base` input to the Terraform module to set the OS base the charm is deployed on.